+++ releases.md
@@ -3,6 +3,7 @@
 
 #### New features
 
+- erf : error function using Abramowitz and Stegun method (maximum error: 1.5e-7) 
 
 #### Bug fixes